Adjust triangulation by by dividing a densely populated triangle into smaller triangles by adding 3 points at the middle of each side of the triangle

Adjust triangulation by by dividing a densely populated triangle into smaller triangles by adding 3 points at the middle of each side of the triangle

vary_density(sample, x, y, label = seq_along(x), basemap = NULL,
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")

Arguments

sample
A data frame containing selected villages for sampling. This is the data frame produced by get_nearest_point() function.

x
A character value specifying the variable name in sample holding the longitude information of the sampling locations.

y
A character value specifying the variable name in sample holding the latitude information of the sampling locations.

label
A character value specifying the variable name in sample holding the name of the sampling locations. If no name information available in data frame, specify a character vector with length equal to number of sampling locations in sample. Default is NULL.

basemap
A SpatialPolygonsDataFrame of base map to show when plotting triangulation. Default is NULL

crs
A character vector specifying the coordinate reference system to use for the spatial objects. Default is longlat projection for datum WGS84

Examples

#